requirements-dev.txt
-r requirements.txt
mypy
pre-commit
pytest
pytest-aiohttp
pytest-asyncio
pytest-cov
types-redis